requirements.txt
cython
numpy
albumentations
imagecorruptions
mmcv==1.5.3
matplotlib
terminaltables
six
pycocotools
pillow==6.2.2
timm
